refactor: trim runtime core scope exceptions
