--- packages/opencode/src/cli/cmd/tui/config/cwd.ts
@@ -1,5 +1,0 @@
-import { Context } from "effect"
-
-export const CurrentWorkingDirectory = Context.Reference<string>("CurrentWorkingDirectory", {
-  defaultValue: () => process.cwd(),
-})
