+++ packages/opencode/src/util/index.ts
@@ -1,3 +1,1 @@
-export { Color } from "./color"
-export { Filesystem } from "./filesystem"
 export { Log } from "./log"
