+++ packages/opencode/test/config/agent-color.test.ts
@@ -5,7 +5,7 @@ import { provideInstance, tmpdir } from "../fixture/fixture"
 import { Instance } from "../../src/project/instance"
 import { Config } from "../../src/config"
 import { Agent as AgentSvc } from "../../src/agent/agent"
-import { Color } from "../../src/util"
+import { Color } from "../../src/util/color"
 import { AppRuntime } from "../../src/effect/app-runtime"
 
 const load = () => AppRuntime.runPromise(Config.Service.use((svc) => svc.get()))
